% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_catboost.R
\name{catboost_predict}
\alias{catboost_predict}
\title{Predict based on the model}
\usage{
catboost_predict(model, pool, prediction_type = "Probability", ...)
}
\arguments{
\item{model}{The trained model.}

\item{pool}{The dataset to predict on.}

\item{prediction_type}{Prediction type.}

\item{...}{Additional parameters.}
}
\value{
Predicted values.
}
\description{
Predict based on the model
}
\examples{
# Example code
}
