% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-catboost.R
\name{cv_catboost}
\alias{cv_catboost}
\title{catboost - parameter tuning and model selection with k-fold cross-validation and grid search}
\usage{
cv_catboost(
  x,
  y,
  nfolds = 5L,
  seed = 42,
  verbose = TRUE,
  iterations = c(10, 50, 100, 200, 500, 1000),
  depth = c(2, 3, 4, 5),
  ncpus = parallel::detectCores()
)
}
\arguments{
\item{x}{Predictor matrix}

\item{y}{Response vector}

\item{nfolds}{Number of folds. Default is 5.}

\item{seed}{Random seed for reproducibility}

\item{verbose}{Show progress?}

\item{iterations}{Grid vector for the parameter \code{iteractions}.}

\item{depth}{Grid vector for the parameter \code{depth}.}

\item{ncpus}{Number of CPU cores to use. Defaults is all detectable cores.}
}
\value{
A data frame containing the complete tuning grid and the AUC values,
with the best parameter combination and the highest AUC value.
}
\description{
catboost - parameter tuning and model selection with k-fold cross-validation and grid search
}
\examples{
# check the vignette for code examples
}
