% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed_catboost}
\alias{is_installed_catboost}
\title{Is catboost installed?}
\usage{
is_installed_catboost()
}
\value{
\code{TRUE} if installed, \code{FALSE} if not.
}
\description{
Is catboost installed?
}
\examples{
is_installed_catboost()
}
