% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed_lightgbm}
\alias{is_installed_lightgbm}
\title{Is lightgbm installed?}
\usage{
is_installed_lightgbm()
}
\value{
\code{TRUE} if installed, \code{FALSE} if not.
}
\description{
Is lightgbm installed?
}
\examples{
is_installed_lightgbm()
}
