% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed_xgboost}
\alias{is_installed_xgboost}
\title{Is xgboost installed?}
\usage{
is_installed_xgboost()
}
\value{
\code{TRUE} if installed, \code{FALSE} if not.
}
\description{
Is xgboost installed?
}
\examples{
is_installed_xgboost()
}
