% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_lightgbm.R
\name{lightgbm_train}
\alias{lightgbm_train}
\title{Train lightgbm model}
\usage{
lightgbm_train(data, label, params, ...)
}
\arguments{
\item{data}{Training data.}

\item{label}{Labels.}

\item{params}{A list of parameters.}

\item{...}{Additional parameters.}
}
\value{
A model object.
}
\description{
Train lightgbm model
}
\examples{
\dontshow{if (is_installed_lightgbm()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example code
\dontshow{\}) # examplesIf}
}
