% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.stackgbm}
\alias{predict.stackgbm}
\title{Make predictions from a stackgbm model object}
\usage{
\method{predict}{stackgbm}(object, newx, threshold = 0.5, classes = c(1L, 0L), ...)
}
\arguments{
\item{object}{stackgbm model object}

\item{newx}{New predictor matrix}

\item{threshold}{Decision threshold. Default is 0.5.}

\item{classes}{The class encoding vector of the predicted outcome.
The naming and order will be respected.}

\item{...}{unused}
}
\value{
A list of two vectors presenting the predicted classification
probabilities and predicted response.
}
\description{
Make predictions from a stackgbm model object
}
\examples{
\dontshow{if (is_installed_xgboost() && is_installed_lightgbm() && is_installed_catboost()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Check the vignette for code examples
\dontshow{\}) # examplesIf}
}
