% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackgbm.R
\name{stackgbm}
\alias{stackgbm}
\title{Model stacking for boosted trees}
\usage{
stackgbm(x, y, params, nfolds = 5L, seed = 42, verbose = TRUE)
}
\arguments{
\item{x}{Predictor matrix}

\item{y}{Response vector}

\item{params}{A list of optimal parameters for boosted tree models.
Can be derived from [cv_xgboost()], [cv_lightgbm()], and [cv_catboost()].}

\item{nfolds}{Number of folds. Default is 5.}

\item{seed}{Random seed for reproducibility}

\item{verbose}{Show progress?}
}
\value{
Fitted boosted tree models and stacked tree model.
}
\description{
Model stacking with a two-layer architecture: first layer being boosted
tree models fitted by xgboost, lightgbm, and catboost; second layer being
a logistic regression model.
}
\examples{
# check the vignette for code examples
}
