% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_xgboost.R
\name{xgboost_dmatrix}
\alias{xgboost_dmatrix}
\title{Create xgb.DMatrix object}
\usage{
xgboost_dmatrix(data, label = NULL, ...)
}
\arguments{
\item{data}{Matrix or file.}

\item{label}{Labels (optional).}

\item{...}{Additional parameters.}
}
\value{
An `xgb.DMatrix` object.
}
\description{
Create xgb.DMatrix object
}
\examples{
# Example code
}
