% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_xgboost.R
\name{xgboost_train}
\alias{xgboost_train}
\title{Train xgboost model}
\usage{
xgboost_train(params, data, nrounds, ...)
}
\arguments{
\item{params}{A list of parameters.}

\item{data}{Training data.}

\item{nrounds}{The Maximum number of boosting iterations.}

\item{...}{Additional parameters.}
}
\value{
A model object.
}
\description{
Train xgboost model
}
\examples{
# Example code
}
