% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_catboost.R
\name{catboost_load_pool}
\alias{catboost_load_pool}
\title{Create a dataset}
\usage{
catboost_load_pool(data, label = NULL, ...)
}
\arguments{
\item{data}{Predictors.}

\item{label}{Labels.}

\item{...}{Additional parameters.}
}
\value{
A `catboost.Pool` object.
}
\description{
Create a dataset
}
\examples{
# Example code
}
