% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_catboost.R
\name{catboost_train}
\alias{catboost_train}
\title{Train the model}
\usage{
catboost_train(learn_pool, test_pool = NULL, params = list())
}
\arguments{
\item{learn_pool}{Training dataset.}

\item{test_pool}{Testing dataset.}

\item{params}{A list of training parameters.}
}
\value{
A model object.
}
\description{
Train the model
}
\examples{
\dontshow{if (is_installed_catboost()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example code
\dontshow{\}) # examplesIf}
}
