% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_catboost.R
\name{cv_catboost}
\alias{cv_catboost}
\title{catboost - parameter tuning and model selection with k-fold cross-validation
and grid search}
\usage{
cv_catboost(
  x,
  y,
  params = cv_param_grid(),
  n_folds = 5,
  n_threads = 1,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Response vector.}

\item{params}{Parameter grid generated by \code{\link[=cv_param_grid]{cv_param_grid()}}.}

\item{n_folds}{Number of folds. Default is 5.}

\item{n_threads}{The number of parallel threads. For optimal speed,
match this to the number of physical CPU cores, not threads.
See respective model documentation for more details. Default is 1.}

\item{seed}{Random seed for reproducibility.}

\item{verbose}{Show progress?}
}
\value{
A data frame containing the complete tuning grid and the AUC values,
with the best parameter combination and the highest AUC value.
}
\description{
catboost - parameter tuning and model selection with k-fold cross-validation
and grid search
}
\examples{
# check the vignette for code examples
}
