% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_xgboost.R
\name{cv_xgboost}
\alias{cv_xgboost}
\title{xgboost - parameter tuning and model selection with k-fold cross-validation
and grid search}
\usage{
cv_xgboost(
  x,
  y,
  params = cv_param_grid(),
  nfolds = 5L,
  seed = 42,
  verbose = TRUE,
  ncpus = parallel::detectCores()
)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Response vector.}

\item{params}{Parameter grid generated by \code{\link[=cv_param_grid]{cv_param_grid()}}.}

\item{nfolds}{Number of folds. Default is 5.}

\item{seed}{Random seed for reproducibility.}

\item{verbose}{Show progress?}

\item{ncpus}{Number of CPU cores to use. Defaults is all detectable cores.}
}
\value{
A data frame containing the complete tuning grid and the AUC values,
with the best parameter combination and the highest AUC value.
}
\description{
xgboost - parameter tuning and model selection with k-fold cross-validation
and grid search
}
\examples{
# check the vignette for code examples
}
