% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_xgboost.R
\name{cv_xgboost}
\alias{cv_xgboost}
\title{xgboost - parameter tuning and model selection with k-fold cross-validation
and grid search}
\usage{
cv_xgboost(
  x,
  y,
  params = cv_param_grid(),
  n_folds = 5,
  n_threads = 1,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Response vector.}

\item{params}{Parameter grid generated by \code{\link[=cv_param_grid]{cv_param_grid()}}.}

\item{n_folds}{Number of folds. Default is 5.}

\item{n_threads}{The number of parallel threads. For optimal speed,
match this to the number of physical CPU cores, not threads.
See respective model documentation for more details. Default is 1.}

\item{seed}{Random seed for reproducibility.}

\item{verbose}{Show progress?}
}
\value{
A data frame containing the complete tuning grid and the AUC values,
with the best parameter combination and the highest AUC value.
}
\description{
xgboost - parameter tuning and model selection with k-fold cross-validation
and grid search
}
\examples{
\dontshow{if (is_installed_xgboost()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Check the vignette for code examples
\dontshow{\}) # examplesIf}
}
