% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackgbm-package.R
\docType{package}
\name{stackgbm-package}
\alias{stackgbm-package}
\alias{_PACKAGE}
\title{stackgbm: Stacked Gradient Boosting Machines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A minimalist implementation of model stacking for boosted tree models built by 'xgboost', 'lightgbm', and 'catboost'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/stackgbm/}
  \item \url{https://github.com/nanxstats/stackgbm}
  \item Report bugs at \url{https://github.com/nanxstats/stackgbm/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

}
\keyword{internal}
